UI polish: ruler sizing, scroll/zoom behavior, drag-drop foundation
Major changes:
- Title bar menu text vertical alignment fix
- FL Studio-style ruler with major bars (14px) vs minor bars (11px)
- Smooth zoom animation with mouse pivot (Shift+scroll to zoom)
- Regular scroll for vertical track movement
- Platform modifier query API for wheel events
- FileBrowser effectiveWidth_ member for consistent bounds
- Disabled preview panel to fix caching glitches
- Drag-drop system foundation (NUIDragDrop.cpp/h)
- TrackUIComponent waveform display improvements

Files modified:
- NomadPlat: getModifiers() API for runtime key state
- NomadUI: DragDrop system, platform bridge modifiers support
- Source: TrackManagerUI scroll/zoom, FileBrowser fixes

+++ NomadAudio/test/WavLoaderTest.cpp
@@ -12,7 +12,7 @@ namespace fs = std::filesystem;
 namespace Nomad {
 namespace Audio {
     bool loadWavFile(const std::string& filePath, std::vector<float>& audioData,
-                     uint32_t& sampleRate, uint16_t& numChannels);
+                     uint32_t& sampleRate, uint32_t& numChannels);
 }
 }
 
@@ -105,7 +105,7 @@ bool runBasic16BitTest() {
 
     std::vector<float> audio;
     uint32_t sampleRate = 0;
-    uint16_t channels = 0;
+    uint32_t channels = 0;
     bool ok = loadWavFile(path, audio, sampleRate, channels);
     fs::remove(path);
 
@@ -131,7 +131,7 @@ bool runJunkChunkTest() {
 
     std::vector<float> audio;
     uint32_t sampleRate = 0;
-    uint16_t channels = 0;
+    uint32_t channels = 0;
     bool ok = loadWavFile(path, audio, sampleRate, channels);
     fs::remove(path);
 
@@ -152,7 +152,7 @@ bool run24BitTest() {
 
     std::vector<float> audio;
     uint32_t sampleRate = 0;
-    uint16_t channels = 0;
+    uint32_t channels = 0;
     bool ok = loadWavFile(path, audio, sampleRate, channels);
     fs::remove(path);
